requirements.txt
streamlit==0.89.0
pandas==1.3.3
together
twilio==6.61.0
google-cloud-speech==2.7.0
transformers==4.11.3
torch==1.9.1
python-dotenv==0.19.1
requests-html==0.10.0
lxml==4.6.3
numpy
